+++ exportplan/views.py
@@ -61,9 +61,16 @@ class ExportPlanTargetMargetsView(ExportPlanSectionView):
             lastyear_import_data = helpers.get_comtrade_lastyearimportdata(
                 commodity_code=commodity_code, country=country
             )
+
+        export_plan = helpers.get_exportplan(sso_session_id=self.request.user.session_id)
+
+        if export_plan:
+            timezone = helpers.get_timezone(export_plan['rules_regulations']['country_code'])
+            utz_offset = datetime.now(pytz.timezone(timezone)).strftime('%z')
+
             return super().get_context_data(
-                rules_regulation=rules_regulation,
-                export_marketdata=export_marketdata,
+                target_markets=export_plan['target_markets'],
+                timezone=timezone,
                 datenow=datetime.now(),
                 utz_offset=utz_offset,
                 lastyear_import_data=lastyear_import_data,
@@ -102,6 +109,7 @@ class ExportPlanStartView(FormView):
             'export_countries': [exportplan_data['country']],
             'export_commodity_codes': [exportplan_data['commodity_code']],
             'rules_regulations': exportplan_data,
+            'target_markets': [{'country': exportplan_data['country']}]
         }